rupees migration: Float -> Numeric(10,2) for wallet/settlement/offer/reward money columns
Converts VendorWallet, VendorTransaction, VendorSettlement, DiscountCampaign,
VendorOffer, and RedemptionRule money columns from Float to Numeric(10,2),
matching the rest of the money model and removing binary-rounding drift risk
on wallet balances that accumulate over many transactions. Alembic migration
20260708_0044 (Postgres-only, no-op on SQLite, precision-only cast — no
data rescaling needed since these were already correctly rupee-scaled).

Fixed 2 real TypeError crash sites this surfaced (Decimal * float literal):
retention_service.py redeem_offer (discount * 0.5 bonus-points calc) and
ranking_service.py rank_offers (avg_affinity * discount_val). Also wrapped
~15 dict-literal HTTP response sites across retention_service.py,
retention_router.py, ranking_service.py, settlement_service.py, and
admin/router.py in explicit float(...) to avoid the known Pydantic
Decimal-to-string serialization bug on routes with loose return-type
annotations.

Verified via 3 real Decimal-valued smoke tests (get_offers, rank_offers,
redeem_offer) plus a full-suite worktree diff against the pre-change commit:
identical 128-failure set, 969 passed, 16 skipped on both sides.

+++ tnt-backend-main/app/modules/rewards/model.py
@@ -82,7 +82,7 @@ class RedemptionRule(Base):
     redemption_type = Column(Enum(RedemptionType, values_callable=lambda x: [e.value for e in x]), nullable=False, unique=True)
     min_points = Column(Float, nullable=False)
     max_discount_percentage = Column(Float, nullable=True)  # For percentage discounts
-    max_discount_amount = Column(Float, nullable=True)  # For fixed discounts
+    max_discount_amount = Column(Numeric(10, 2), nullable=True)  # For fixed discounts
     is_active = Column(Integer, nullable=False, default=1)
     created_at = Column(DateTime, default=utcnow_naive)
     updated_at = Column(DateTime, default=utcnow_naive, onupdate=utcnow_naive)
